feat: added both p2p and normal data models on issue tab
