+++ apps/website/components/templates/explore/tabs/all-tab/banner/card-earn-credential.tsx
@@ -23,7 +23,7 @@ export function CardEarnCredential(): JSX.Element {
         display: 'flex',
         justifyContent: 'space-between',
         borderRadius: '16px',
-        ml: '18px',
+        ml: '25px',
       }}
     >
       <Stack
